Changes table width and color styling, reorganizes functions into modules
