+++ src/js/index.js
@@ -1,6 +1,5 @@
-import React, { Component } from 'react';
-import ReactDOM from 'react-dom';
-import AppComponent from './AppComponent.js';
+import React, { Component } from "react";
+import ReactDOM from "react-dom";
+import AppComponent from "./AppComponent.js";
 
-
-ReactDOM.render(<AppComponent />, document.getElementById('root'));
+ReactDOM.render(<AppComponent />, document.getElementById("root"));
